+++ app/models/document.rb
@@ -1,4 +1,4 @@
 class Document < ApplicationRecord
   validates :title, presence: true
-  validates :slug, presence: true
+  validates :slug, presence: true, uniqueness: true
 end